real_time_pipeline/requirements.txt
yfinance
pandas
google-cloud-pubsub